refactorings re: business structure
