--- test/resources/tables/business_structure.csv
@@ -1,15 +1,0 @@
-workerVAT,businessAccount,advertiseForWork,businessWebsite,workerPayForTraining,workerExpenseRunningBusinessPremises,workerPaysForInsurance,CarryOverValue ,Exit,CarryOverFactName
-Yes,Yes,Yes,Yes,Yes,Yes,Yes,High,FALSE,
-Yes,Yes,Yes,Yes,Yes,Yes,No,High,FALSE,
-Yes,Yes,Yes,Yes,Yes,No,No,High,FALSE,
-Yes,Yes,Yes,Yes,No,No,No,High,FALSE,
-Yes,Yes,Yes,No,No,No,No,Low,FALSE,
-Yes,Yes,No,No,No,No,No,Low,FALSE,
-Yes,No,No,No,No,No,No,Low,FALSE,
-No,No,No,No,No,No,No,Low,FALSE,
-No,No,No,No,No,No,Yes,Low,FALSE,
-No,No,No,No,No,Yes,Yes,Low,FALSE,
-No,No,No,No,Yes,Yes,Yes,Low,FALSE,
-No,No,No,Yes,Yes,Yes,Yes,High,FALSE,
-No,No,Yes,Yes,Yes,Yes,Yes,High,FALSE,
-No,Yes,Yes,Yes,Yes,Yes,Yes,High,FALSE,
