+++ test/uk/gov/hmrc/decisionservice/controllers/DecisionControllerSpec.scala
@@ -28,7 +28,7 @@ import uk.gov.hmrc.decisionservice.model.FactError
 import uk.gov.hmrc.decisionservice.model.api.{DecisionRequest, Score}
 import uk.gov.hmrc.decisionservice.model.rules.Facts
 import uk.gov.hmrc.decisionservice.ruleengine.{RuleEngineDecision, RulesFileMetaData}
-import uk.gov.hmrc.decisionservice.services.DecisionService
+import uk.gov.hmrc.decisionservice.services.{DecisionService, DecisionServiceTestInstance}
 import uk.gov.hmrc.play.test.{UnitSpec, WithFakeApplication}
 
 class DecisionControllerSpec extends UnitSpec with WithFakeApplication {
@@ -44,16 +44,9 @@ class DecisionControllerSpec extends UnitSpec with WithFakeApplication {
     (24, "/tables/financial_risk.csv", "financial_risk"),
     (5,  "/tables/part_of_organisation.csv", "part_of_organisation"),
     (1,  "/tables/misc.csv", "miscellaneous"),
-    (7,  "/tables/business_structure.csv", "business_structure"),
     (13, "/tables/personal_service.csv", "personal_service"),
     (6,  "/tables/matrix_of_matrices.csv", "matrix")
   ).collect{case (q,f,n) => RulesFileMetaData(q,f,n)}
-
-
-  object DecisionServiceTestInstance extends DecisionService {
-    lazy val maybeSectionRules = loadSectionRules()
-    lazy val csvSectionMetadata = testCsvSectionMetadata
-  }
 
   object ErrorGeneratingDecisionService extends DecisionService {
     lazy val maybeSectionRules = loadSectionRules()
