+++ test/uk/gov/hmrc/decisionservice/services/ComplianceTestCasesSpec.scala
@@ -18,21 +18,10 @@ package uk.gov.hmrc.decisionservice.services
 
 import uk.gov.hmrc.decisionservice.model.api.Score
 import uk.gov.hmrc.decisionservice.model.rules.{EmptyCarryOver, _}
+import uk.gov.hmrc.decisionservice.ruleengine.FactValidatingFunctions._
+import uk.gov.hmrc.decisionservice.ruleengine.MatchingFunctions._
 import uk.gov.hmrc.decisionservice.ruleengine.RulesFileMetaData
 import uk.gov.hmrc.play.test.UnitSpec
-
-object DecisionServiceTestInstance extends DecisionService {
-  lazy val maybeSectionRules = loadSectionRules()
-  lazy val csvSectionMetadata = List(
-    (13, "/tables/control.csv", "control"),
-    (24, "/tables/financial_risk.csv", "financial_risk"),
-    (5, "/tables/part_of_organisation.csv", "part_of_organisation"),
-    (1, "/tables/misc.csv", "miscellaneous"),
-    (7, "/tables/business_structure.csv", "business_structure"),
-    (13, "/tables/personal_service.csv", "personal_service"),
-    (6, "/tables/matrix_of_matrices.csv", "matrix")
-  ).collect { case (q, f, n) => RulesFileMetaData(q, f, n) }
-}
 
 class ComplianceTestCasesSpec extends UnitSpec {
   val testCases = Map(
@@ -65,7 +54,7 @@ class ComplianceTestCasesSpec extends UnitSpec {
           val maybeBusinessStructureScore = Score.create(decision.facts).get("business_structure")
           maybeBusinessStructureScore.isDefined shouldBe true
           maybeBusinessStructureScore.map{businessStructureScore =>
-            businessStructureScore shouldBe NotValidUseCase.value
+            businessStructureScore shouldBe "low"
           }
         }
       }
