+++ test/uk/gov/hmrc/decisionservice/services/TablesFinalDecisionSpec.scala
@@ -30,7 +30,6 @@ class TablesFinalDecisionSpec extends UnitSpec {
       (24, "/tables/financial_risk.csv", "financial_risk"),
       (5,  "/tables/part_of_organisation.csv", "part_of_organisation"),
       (1,  "/tables/misc.csv", "miscellaneous"),
-      (7,  "/tables/business_structure.csv", "business_structure"),
       (13, "/tables/personal_service.csv", "personal_service"),
       (6,  "/tables/matrix_of_matrices.csv", "matrix")
     ).collect{case (q,f,n) => RulesFileMetaData(q,f,n)}
